moved the logic of saving session results to another class
