+++ maven-file/src/main/resources/Count of unique words.txt
@@ -1,3 +1,3 @@
-default text is Some text and other something interesting some
-[ text ][ and ][ other ][ something ][ interesting ][ some ] 
- Text has 6 unique word
+default text is dfdf
+[ dfdf ]
+ Text has 1 unique word
