+++ maven-file/src/main/resources/File with information about count of matches.txt
@@ -1,3 +1,3 @@
-default text is Some text and other something interesting some
- Searchable word is sdsd
+default text is dfdf
+ Searchable word is sd
 Matches count is 0
