+++ maven-file/src/main/resources/File with up-case letters.txt
@@ -1,4 +1,4 @@
-default text is Some text and other something interesting some
+default text is dfdf
 
-Current text is S O M E T E X T A N D O T H E R S O M E T H I N G I N T E R E S T I N G S O M E
-Count of letters at whole is 40
+Current text is D F D F
+Count of letters at whole is 4
