+++ maven-file/src/main/resources/Result.txt
@@ -1,1 +1,18 @@
 =========
+default text is dfdf
+[ dfdf ]
+ Text has 1 unique word
+
+
+default text is dfdf
+
+Current text is D F D F
+Count of letters at whole is 4
+
+
+default text is dfdf
+ Searchable word is sd
+Matches count is 0
+
+
+===============
